+++ src/main/java/com/emeraldElves/alcohollabelproject/Data/IAssigner.java
@@ -1,4 +1,4 @@
-package com.emeraldElves.alcohollabelproject.UserInterface;
+package com.emeraldElves.alcohollabelproject.Data;
 
 import java.util.List;
