+++ src/main/java/com/emeraldElves/alcohollabelproject/Data/SimpleApplicationAssigner.java
@@ -1,4 +1,7 @@
-package com.emeraldElves.alcohollabelproject;
+package com.emeraldElves.alcohollabelproject.Data;
+
+import com.emeraldElves.alcohollabelproject.Data.ApplicationAssigner;
+import com.emeraldElves.alcohollabelproject.Data.RoundRobinAssigner;
 
 import java.util.List;
